+++ packages/liveblocks-core/src/store.ts
@@ -33,6 +33,8 @@ type OptimisticUpdate<M extends BaseMetadata> =
   | RemoveReactionOptimisticUpdate
   | MarkInboxNotificationAsReadOptimisticUpdate
   | MarkAllInboxNotificationsAsReadOptimisticUpdate
+  | DeleteInboxNotificationOptimisticUpdate
+  | DeleteAllInboxNotificationsOptimisticUpdate
   | UpdateNotificationSettingsOptimisticUpdate;
 
 type CreateThreadOptimisticUpdate<M extends BaseMetadata> = {
@@ -119,9 +121,22 @@ type MarkInboxNotificationAsReadOptimisticUpdate = {
 };
 
 type MarkAllInboxNotificationsAsReadOptimisticUpdate = {
-  type: "mark-inbox-notifications-as-read";
+  type: "mark-all-inbox-notifications-as-read";
   id: string;
   readAt: Date;
+};
+
+type DeleteInboxNotificationOptimisticUpdate = {
+  type: "delete-inbox-notification";
+  id: string;
+  inboxNotificationId: string;
+  deletedAt: Date;
+};
+
+type DeleteAllInboxNotificationsOptimisticUpdate = {
+  type: "delete-all-inbox-notifications";
+  id: string;
+  deletedAt: Date;
 };
 
 type UpdateNotificationSettingsOptimisticUpdate = {
@@ -565,13 +580,25 @@ export function applyOptimisticUpdates<M extends BaseMetadata>(
         };
         break;
       }
-      case "mark-inbox-notifications-as-read": {
+      case "mark-all-inbox-notifications-as-read": {
         for (const id in result.inboxNotifications) {
           result.inboxNotifications[id] = {
             ...result.inboxNotifications[id],
             readAt: optimisticUpdate.readAt,
           };
         }
+        break;
+      }
+      case "delete-inbox-notification": {
+        const {
+          [optimisticUpdate.inboxNotificationId]: _,
+          ...inboxNotifications
+        } = result.inboxNotifications;
+        result.inboxNotifications = inboxNotifications;
+        break;
+      }
+      case "delete-all-inbox-notifications": {
+        result.inboxNotifications = {};
         break;
       }
       case "update-notification-settings": {