+++ packages/liveblocks-react/src/__tests__/useUnreadInboxNotificationsCount.test.tsx
@@ -1,16 +1,14 @@
 import "@testing-library/jest-dom";
 
-import { createClient } from "@liveblocks/core";
 import { renderHook, waitFor } from "@testing-library/react";
 import { setupServer } from "msw/node";
 import { nanoid } from "nanoid";
 import React, { Suspense } from "react";
 
-import { createLiveblocksContext } from "../liveblocks";
 import { dummyThreadData, dummyThreadInboxNotificationData } from "./_dummies";
 import MockWebSocket from "./_MockWebSocket";
 import { mockGetInboxNotifications } from "./_restMocks";
-import { generateFakeJwt } from "./_utils";
+import { createContextsForTest } from "./_utils";
 
 const server = setupServer();
 
@@ -26,22 +24,6 @@ afterEach(() => {
 });
 
 afterAll(() => server.close());
-
-// TODO: Dry up and create utils that wrap renderHook
-function createLiveblocksContextForTest() {
-  const client = createClient({
-    async authEndpoint() {
-      return {
-        token: await generateFakeJwt({ userId: "userId" }),
-      };
-    },
-    polyfills: {
-      WebSocket: MockWebSocket as any,
-    },
-  });
-
-  return createLiveblocksContext(client);
-}
 
 describe("useUnreadInboxNotificationsCount", () => {
   test("should fetch inbox notification count", async () => {
@@ -70,8 +52,9 @@ describe("useUnreadInboxNotificationsCount", () => {
       })
     );
 
-    const { LiveblocksProvider, useUnreadInboxNotificationsCount } =
-      createLiveblocksContextForTest();
+    const {
+      liveblocks: { LiveblocksProvider, useUnreadInboxNotificationsCount },
+    } = createContextsForTest();
 
     const { result, unmount } = renderHook(
       () => useUnreadInboxNotificationsCount(),
@@ -125,8 +108,10 @@ describe("useUnreadInboxNotificationsCount - Suspense", () => {
     );
 
     const {
-      suspense: { LiveblocksProvider, useUnreadInboxNotificationsCount },
-    } = createLiveblocksContextForTest();
+      liveblocks: {
+        suspense: { LiveblocksProvider, useUnreadInboxNotificationsCount },
+      },
+    } = createContextsForTest();
 
     const { result, unmount, rerender } = renderHook(
       () => useUnreadInboxNotificationsCount(),