+++ packages/liveblocks-react/src/index.ts
@@ -24,6 +24,8 @@ export {
   useInboxNotificationThread,
   useMarkAllInboxNotificationsAsRead,
   useMarkInboxNotificationAsRead,
+  useDeleteAllInboxNotifications,
+  useDeleteInboxNotification,
 } from "./liveblocks";
 export {
   createRoomContext,